+++ tests/build_tests/duplicated_symlinked_packages/input.js
@@ -12,7 +12,14 @@ function postProcessErrorOutput(output) {
   output = output.replace(new RegExp(__dirname, "gi"), ".");
   return output;
 }
+
+if (process.platform === "win32") {
+  console.log("Skipping test on Windows");
+  process.exit(0);
+}
+
 child_process.execSync(`${rescript_exe} clean`, { cwd: __dirname });
+
 child_process.exec(rescript_exe, { cwd: __dirname }, (err, stdout, stderr) => {
   const actualErrorOutput = postProcessErrorOutput(stderr.toString());
   if (updateTests) {