+++ tests/build_tests/unicode/input.js
@@ -1,11 +1,17 @@
 //@ts-check
 var child_process = require("child_process");
+var fs = require("fs");
+var path = require("path");
+
 var { rescript_exe } = require("#cli/bin_path");
+
+if (process.platform === "win32") {
+  console.log("Skipping test on Windows");
+  process.exit(0);
+}
 
 console.log(child_process.execSync(rescript_exe, { encoding: "utf8" }));
 
-var fs = require("fs");
-var path = require("path");
 var content =
   "" + fs.readFileSync(path.join(__dirname, "lib", "bs", ".sourcedirs.json"));