+++ tests/build_tests/weird_deps/input.js
@@ -3,6 +3,7 @@
 var cp = require("child_process");
 var assert = require("assert");
 var { rescript_exe } = require("#cli/bin_path");
+var { normalizeNewlines } = require("../utils.js");
 
 var out = cp.spawnSync(rescript_exe, {
   cwd: __dirname,
@@ -13,7 +14,7 @@ if (out.stdout !== "") {
   assert.fail(out.stdout);
 } else {
   assert.equal(
-    out.stderr,
+    normalizeNewlines(out.stderr),
     [
       'File "bsconfig.json", line 1',
       "Error: package weird not found or built ",