+++ CHANGELOG.md
@@ -46,6 +46,7 @@
 - Remove uncurried flag from bsb. https://github.com/rescript-lang/rescript-compiler/pull/7049
 - Build runtime/stdlib files with rescript/bsb instead of ninja.js. https://github.com/rescript-lang/rescript-compiler/pull/7063
 - Build tests with bsb and move them out of jscomp. https://github.com/rescript-lang/rescript-compiler/pull/7068
+- Run `build_tests` on Windows. https://github.com/rescript-lang/rescript-compiler/pull/7065
 
 # 12.0.0-alpha.3